requirements.txt
tkinter
pillow
rsa
onetimepad
optparse
fernet